backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3
django-prometheus==2.3.1
prometheus-client==0.20.0
